+++ src/org/noisesmith/philoseek.clj
@@ -7,6 +7,14 @@
   [& tags]
   (filter (comp (set tags) :tag)))
 
+(defn not-internal-resource?
+  [uri]
+  (every? (fn [s] (not (string/starts-with? uri s)))
+          ["/wiki/Wikipedia:"
+           "/wiki/File:"
+           "/wiki/Help:"
+           "/wiki/Talk:"]))
+
 (def find-wiki
   "based on a bunch of repl exploration using the dumped data, this seems
    to be the thing that will predictably get us the right link"
@@ -16,11 +24,14 @@
                     (:href)
                     (as-> uri
                       (and (string/starts-with? uri "/wiki/")
-                           (not (string/starts-with? uri "/wiki/Wikipedia:"))))))))
+                           (not-internal-resource? uri)))))))
+
+(def debug (atom nil))
 
 (defn extract-link
   "function that exists for exploring our data in the repl"
    [body]
+   (reset! debug body)
     (-> body
         (->> (tree-seq coll? seq)
              (sequence (comp (find-tag :a)
@@ -31,8 +42,9 @@
 
 (defn raw-search
   "Tries to find the Philosophy page on wikipedia in a roundabout manner."
-  ([] (raw-search "https://en.wikipedia.org/wiki/Special:Random"))
+  ([] (raw-search "/wiki/Special:Random"))
   ([url]
-   (-> url
-       (x/parse)
-       (extract-link))))
+   (->> url
+        (str "https://en.wikipedia.org")
+        (x/parse)
+        (extract-link))))